pyproject.toml
[project]
name = "qna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix>=11.0.0",
    "fastapi>=0.115.13",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.24",
    "mcp-use>=1.3.1",
    "nest-asyncio>=1.6.0",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "streamlit>=1.46.0",
]
